Return results from ProfileManager.finalize(return_results=True)
