+++ src/scope_profiler/inspection.py
@@ -15,6 +15,7 @@ from pathlib import Path
 import numpy as np
 
 from scope_profiler.h5reader import ProfilingH5Reader
+from scope_profiler.results import ProfilingResults
 from scope_profiler.summary import SORT_KEYS, print_region_table, region_rows
 
 # Metadata is printed in these groups, in this order, so the fields that
@@ -203,13 +204,13 @@ def _json_safe(value):
 
 
 def collect_file_metadata(
-    profiling_data: ProfilingH5Reader | str | Path | Sequence,
+    profiling_data: ProfilingResults | str | Path | Sequence,
 ) -> dict:
     """Collect the run metadata of one or more profiling files.
 
     Parameters
     ----------
-    profiling_data : ProfilingH5Reader, path, or sequence of either
+    profiling_data : ProfilingResults, path, or sequence of either
         Files to read the metadata from.
 
     Returns
@@ -220,14 +221,12 @@ def collect_file_metadata(
         :func:`~scope_profiler.plotting_scripts.collect_region_statistics`, so
         a single document can describe several runs.
     """
-    if isinstance(profiling_data, (ProfilingH5Reader, str, Path)):
+    if isinstance(profiling_data, (ProfilingResults, str, Path)):
         profiling_data = [profiling_data]
 
     files = []
     for item in profiling_data:
-        reader = (
-            item if isinstance(item, ProfilingH5Reader) else ProfilingH5Reader(item)
-        )
+        reader = item if isinstance(item, ProfilingResults) else ProfilingH5Reader(item)
         files.append(
             {
                 "file_path": str(Path(reader.file_path).resolve()),
@@ -242,14 +241,14 @@ def collect_file_metadata(
 
 
 def write_metadata_json(
-    profiling_data: ProfilingH5Reader | str | Path | Sequence,
+    profiling_data: ProfilingResults | str | Path | Sequence,
     filepath: str | Path,
 ) -> dict:
     """Write the run metadata of one or more profiling files to JSON.
 
     Parameters
     ----------
-    profiling_data : ProfilingH5Reader, path, or sequence of either
+    profiling_data : ProfilingResults, path, or sequence of either
         Files to read the metadata from.
     filepath : str or Path
         Destination JSON file. Parent directories are created as needed.
